+++ game_patch/multi/gametype.cpp
@@ -2021,7 +2021,7 @@ void multi_level_init_post_gametypes()
 // pre level being loaded
 CodeInjection multi_level_init_gametypes_injection{
     0x0046E466,
-    [](auto& regs) {
+    [] {
         rounds_level_init();
         hill_mode_level_init();
         bagman_level_init();